seperated userdata into 2 files
